# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Roisfaozi_queue-base_internal_modules_scanner_model.CheckInRequest:
    properties:
      action:
        enum:
        - register
        - forward
        type: string
      destination_counter_id:
        type: string
      destination_service_id:
        type: string
      patient_id:
        type: string
      patient_name:
        type: string
      queue_id:
        type: string
      service_id:
        type: string
    required:
    - action
    type: object
  github_com_Roisfaozi_queue-base_internal_modules_service_model.CreateServiceRequest:
    properties:
      code:
        maxLength: 50
        minLength: 2
        type: string
      is_pharmacy:
        type: boolean
      is_pharmacy_reception:
        type: boolean
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - code
    - name
    type: object
  github_com_Roisfaozi_queue-base_internal_modules_service_model.UpdateServiceRequest:
    properties:
      code:
        maxLength: 50
        minLength: 2
        type: string
      is_pharmacy:
        type: boolean
      is_pharmacy_reception:
        type: boolean
      name:
        maxLength: 255
        minLength: 3
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.PageMetadata:
    properties:
      limit:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      total_item:
        type: integer
      total_page:
        type: integer
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper:
    properties:
      data: {}
      paging:
        $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.PageMetadata'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a clean and modular boilerplate for Go REST APIs with RBAC,
    Audit Logs, and WebSockets.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Clean Boilerplate API
  version: "1.0"
paths:
...
  /scanner/check-in:
    post:
      consumes:
      - application/json
      description: Scanner entrypoint for queue registration or forward transition
        using client headers.
      parameters:
      - description: Scanner client ID
        in: header
        name: X-Client-ID
        required: true
        type: string
      - description: Scanner API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Scanner payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_scanner_model.CheckInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Scanner check-in
      tags:
      - scanner
  /services:
    get:
      description: Returns all services under active tenant scope.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: List services
      tags:
      - services
    post:
      consumes:
      - application/json
      description: Creates a new service under active tenant scope.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Create Service Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_service_model.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Create service
      tags:
      - services
  /services/{id}:
    delete:
      description: Deletes service under active tenant scope.
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Delete service
      tags:
      - services
    get:
      description: Returns service details under active tenant scope.
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get service by ID
      tags:
      - services
    put:
      consumes:
      - application/json
      description: Updates service fields under active tenant scope.
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Update Service Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_service_model.UpdateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Update service
      tags:
      - services
...
securityDefinitions:
  BearerAuth:
    description: '"Type ''Bearer '' followed by a space and the access token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"